\name{calc_z_var}
\alias{calc_z_var}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal mGSZ function}
\description{Internal mGSZ function not to be called by the users}
\author{Pashupati Mishra, Petri Toronen}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
