\name{count_hyge_var_mean}
\alias{count_hyge_var_mean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal mGSZ function}
\description{Internal mGSZ functions not to be called by the users}
\author{Pashupati Mishra, Petri Toronen}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
