\name{geneSetsList}
\alias{geneSetsList}
\title{Convert gene set data in gmt file to R list}
\description{Converts gene set data in gmt file to R list readable by mGSZ program}
\usage{
geneSetsList(data)
}

\arguments{
  \item{data}{Gene set data in gmt file format
}
}

\value{
Gene set data in list format with gene set name as list names}
\author{Pashupati Mishra, Petri Toronen}

\examples{
## gene.sets <- geneSetsList(filename.gmt) ##
}
