\name{KS.score}
\alias{KS.score}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal KS function}
\description{Internal KS function not to be called by the users}

\author{Pashupati Mishra, Petri Toronen}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
