\name{mGSZ.test.score}
\alias{mGSZ.test.score}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal mGSZ function}
\description{Internal mGSZ function not to be called by the users}
\author{Pashupati Mishra, Petri Toronen}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line